Add Depth Option
The depth dictates how far to descend in to the directory tree. All files
are added to the process list but only folders at the provided depth are
added. The general idea is the torrent output folder will be provided as
the input with a depth of 0. If files are in folders the folder needs to
be processed; if it's just a file it will be up a level (at the source)
and that needs to be processed.