Add File and Folder Ignoring
The rules.json file can now contain a list of regexs to ignore files
and folders. If a file is encountered in the ignore list no rule will
be run on it and it won't be written out anywhere (except debug log).
If a folder is encountered it won't be descended into.