Support Named Matches for Command Formatting
This change allows you to have named groups in your regular expressions.
Those named groups are then passed back to the command formatting for the
matches so you can make use of the values in the matched file. For example,
you can match the year and month on "The.Daily.Show.2014.11.17..." to
ensure the year and month folder exist when moving that file.

Usage:
Used the (?P<some_name>...) style named groups. In command for that match,
you would use something like
echo {some_name}